Renaming of NewTeamReturn and Team.
Team is now known as ReceivedTeam.
NewTeamReturn is now known as Team.
